- move cart logic from  to hook - fix products population error
